fix: 작성페이지 상태 여부 체크하는 로직 다시 stepMenuItem -> layout 으로 이동 (버그로 인해)
